+++ website/pelicanconf.py
@@ -33,14 +33,19 @@ DISPLAY_PAGES_ON_MENU = False
 CATEGORY_URL = "articles/"
 CATEGORY_SAVE_AS = "articles/index.html"
 
+# Disable the default archives page
+ARCHIVES_SAVE_AS = ""
+
 FILENAME_METADATA = r"(?P<date>\d{4}-\d{2}-\d{2})-(?P<slug>.+)"
 
 THEME = "./theme/"
 STATIC_PATHS = [
     "../../brand/favicon.ico",
     "../../brand/dragonfly-rainbow.svg",
+    "../archive-redirect.html",
 ]
 EXTRA_PATH_METADATA = {k: {"path": basename(k)} for k in STATIC_PATHS}
+EXTRA_PATH_METADATA["../archive-redirect.html"] = {"path": "archives.html"}
 PROFILE_IMAGE_URL = "/dragonfly-rainbow.svg"
 
 MENUITEMS = (
@@ -49,6 +54,34 @@ MENUITEMS = (
     ("GitHub", "https://github.com/HypothesisWorks/hypothesis/"),
     ("PyPI", "https://pypi.org/project/hypothesis/"),
 )
+
+# Author information - map from short alias to full name and URL
+AUTHOR_NAMES = {
+    "alexwlchan": "Alex Chan",
+    "carlini": "Nicholas Carlini",
+    "drmaciver": "David R. MacIver",
+    "giorgiosironi": "Giorgio Sironi",
+    "hwayne": "Hillel Wayne",
+    "jml": "Jonathan M. Lange",
+    "liam": "Liam DeVoe",
+    "maaz": "Muhammad Maaz",
+    "nchammas": "Nicholas Chammas",
+    "zac-hd": "Zac Hatfield-Dodds",
+}
+
+AUTHOR_URLS = {
+    "alexwlchan": "https://alexwlchan.net",
+    "carlini": "https://nicholas.carlini.com/",
+    "drmaciver": "http://www.drmaciver.com",
+    "giorgiosironi": "http://giorgiosironi.com",
+    "hwayne": "https://www.hillelwayne.com/",
+    "jml": "https://jml.io",
+    "liam": "https://tybug.dev",
+    "maaz": "https://www.mmaaz.ca/",
+    "nchammas": "http://nchammas.com",
+    "zac-hd": "https://zhd.dev",
+}
+assert set(AUTHOR_URLS).issubset(AUTHOR_NAMES)
 
 DEFAULT_PAGINATION = False